src/whattheversion/requirements.txt
pydantic
GitPython
requests
pyyaml